+++ app/components/Footer.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 
 export default function Footer() {
     return (
-        <div className="container">
+        <div className="container-fluid">
             <footer className="py-3 my-4">
                 <ul className="nav justify-content-center border-bottom pb-3 mb-3">
                     <li className="nav-item"><Link href="#" className="nav-link px-2 text-muted">Home</Link></li>
