+++ app/layout.tsx
@@ -7,9 +7,7 @@ export default function RootLayout({children,}: {children: React.ReactNode}) {
       <head>
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
       </head>
-      <body>
-        {children}
-      </body>
+      <body>{children}</body>
     </html>
   )
 }
