+++ package.json
@@ -18,8 +18,8 @@
     "@types/react-dom": "^18.2.7",
     "bootstrap": "^5.3.0",
     "eslint": "8.43.0",
-    "eslint-config-next": "13.4.6",
-    "next": "^13.4.12",
+    "eslint-config-next": "^13.5.3",
+    "next": "^13.5.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.2",
