+++ public/assets/styles/page.module.scss
@@ -3,8 +3,6 @@ $secondary: #8b7e74;
 $tertiary: #c7bca1;
 $tertiary-2: #f1d3b3;
 
-.main {
-}
 
 .Registration {
   color: hsl(217, 10%, 50.8%);
@@ -14,6 +12,9 @@ $tertiary-2: #f1d3b3;
   }
 }
 
+/*--------------
+  # LOGIN START
+----------------*/
 .Login {
   .divider:after,
   .divider:before {
@@ -23,6 +24,13 @@ $tertiary-2: #f1d3b3;
     background: #eee;
   }
 }
+/*--------------
+# LOGIN END
+----------------*/
+
+/*--------------
+  # FORGOT PASSWORD START
+----------------*/
 
 .ForgotPassword {
   .card {
@@ -30,6 +38,13 @@ $tertiary-2: #f1d3b3;
   }
 }
 
+/*--------------
+  # FORGOT PASSWORD  END
+----------------*/
+
+/*-------------------------
+  # NAVBAR SECTION
+---------------------------*/
 .Navbar {
   width: 100%;
   margin: 20px 0;
@@ -58,10 +73,13 @@ $tertiary-2: #f1d3b3;
     background-color: #ffede7;
   }
 }
+/*-------------------------
+  # NAVBAR SECTION END
+---------------------------*/
 
-/*
-  KYC SECTION START
-*/
+/*------------
+  # KYC SECTION START
+-------------*/
 .kycPageSection {
   //background-color: #d9d9d9;
   display: flex;
@@ -78,11 +96,14 @@ $tertiary-2: #f1d3b3;
     float: right;
   }
 }
-/*
-  KYC SECTION END
-*/
+/*----------
+  # KYC SECTION END
+-----------*/
 
-.landingPageSection {
+/*------------------
+ # LANDING PAGE START
+--------------------*/
+.main {
   /*--------------------------------------------------------------
 # Hero Section
 --------------------------------------------------------------*/
@@ -213,3 +234,19 @@ $tertiary-2: #f1d3b3;
     transform: scale(1);
   }
 }
+/*------------------
+ # LANDING PAGE END
+--------------------*/
+
+
+/*------------------
+ # DASHBOARD PAGE START
+--------------------*/
+.dashboardSection {
+   .fundHero {
+    margin-top:30px;
+  }
+}
+/*------------------
+ # DASHBOARD PAGE END
+--------------------*/
